==Timers & More == =General: *Some updates & fixes in ItemDB. -Fixed Ur's Plate DEF & MDEF. -Also fixed STR-Bonus of King Dramoh Card. -Added full support for Birthday IdRO10th Scroll.
*Fixed a linker warning if compiled with VS 2012.

*Changed 'tick' variables to 64 bit
-This fixes an issue with timers that stop working after about 24-49 days when the tick overflows (note that this may happen much earlier than that, and at hard to predict times, on some systems)
-Updated the RDTSC help message in the configure script to also warn users about issues with SpeedStep enabled systems.
-On Windows, tick() still has a resolution of 10~15ms (or even as low as 100ms on some systems).
-Note: on Windows versions earlier than 6.x (Vista, Server 2008), the tick overflow issue is NOT fixed, since they don't support the function used to retrieve a 64 bit tick. This isn't a big issue, since
	those platforms are already - or going soon to be - out of their extended support period, and it's already advisable to upgrade, for other reasons. If you're the unfortunate user of such a system, it is
	recommended that you reboot your machine at least once every 49 days for 15-3athena to work reliably.
-Note: To clear some doubts, this has absolutely NOTHING to do with 32/64 bit CPUs or OSes. It's all about a variable's size, not the size of registers of your CPU, and a 32bit CPU will be able to handle this just
	fine.
-Thanks to @HerculesWS.

*Fixed & updated item_package system.
-Random items will now drop randomly.
-all items are identified.
-added a new flag: isAnnounced.
--Setting this to 1, will broadcast this item if dropped.
--also added full support for ZC_BROADCASTING_SPECIAL_ITEM_OBTAIN. This is needed to broadcast this drop. A 2009-12-01aRagexeRE or newer is required for this packet.

+++ src/char/char.c
@@ -178,7 +178,7 @@ struct online_char_data {
 };
 
 static DBMap* online_char_db; // int account_id -> struct online_char_data*
-static int chardb_waiting_disconnect(int tid, unsigned int tick, int id, intptr_t data);
+static int chardb_waiting_disconnect(int tid, int64 tick, int id, intptr_t data);
 
 static void* create_online_char_data(DBKey key, va_list args)
 {
@@ -1236,7 +1236,7 @@ void mmo_char_sync(void)
 //----------------------------------------------------
 // Function to save (in a periodic way) datas in files
 //----------------------------------------------------
-int mmo_char_sync_timer(int tid, unsigned int tick, int id, intptr_t data)
+int mmo_char_sync_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	if (save_log)
 		ShowInfo("Saving all files...\n");
@@ -2210,7 +2210,7 @@ static void char_auth_ok(int fd, struct char_session_data *sd)
 	// continues when account data is received...
 }
 
-int send_accounts_tologin(int tid, unsigned int tick, int id, intptr_t data);
+int send_accounts_tologin(int tid, int64 tick, int id, intptr_t data);
 void mapif_server_reset(int id);
 
 
@@ -2643,9 +2643,9 @@ int parse_fromlogin(int fd)
 	return 0;
 }
 
-int check_connect_login_server(int tid, unsigned int tick, int id, intptr_t data);
-// int ping_login_server(int tid, unsigned int tick, int id, intptr_t data); //Disabled, no need anymore. [15peaces]
-int send_accounts_tologin(int tid, unsigned int tick, int id, intptr_t data);
+int check_connect_login_server(int tid, int64 tick, int id, intptr_t data);
+// int ping_login_server(int tid, int64 tick, int id, intptr_t data); //Disabled, no need anymore. [15peaces]
+int send_accounts_tologin(int tid, int64 tick, int id, intptr_t data);
 
 void do_init_loginif(void)
 {
@@ -4463,7 +4463,7 @@ int mapif_send(int fd, unsigned char *buf, unsigned int len)
 	return 0;
 }
 
-int broadcast_user_count(int tid, unsigned int tick, int id, intptr_t data)
+int broadcast_user_count(int tid, int64 tick, int id, intptr_t data)
 {
 	uint8 buf[6];
 	int users = count_users();
@@ -4509,7 +4509,7 @@ static int send_accounts_tologin_sub(DBKey key, void* data, va_list ap)
 	return 0;
 }
 
-int send_accounts_tologin(int tid, unsigned int tick, int id, intptr_t data)
+int send_accounts_tologin(int tid, int64 tick, int id, intptr_t data)
 {
 	if (login_fd > 0 && session[login_fd])
 	{
@@ -4527,7 +4527,7 @@ int send_accounts_tologin(int tid, unsigned int tick, int id, intptr_t data)
 	return 0;
 }
 
-int check_connect_login_server(int tid, unsigned int tick, int id, intptr_t data)
+int check_connect_login_server(int tid, int64 tick, int id, intptr_t data)
 {
 	if (login_fd > 0 && session[login_fd] != NULL)
 		return 0;
@@ -4561,7 +4561,7 @@ int check_connect_login_server(int tid, unsigned int tick, int id, intptr_t data
 
 /* // Disabled, no need anymore. [15peaces]
 // sends a ping packet to login server (will receive pong 0x2718)
-int ping_login_server(int tid, unsigned int tick, int id, intptr_t data)
+int ping_login_server(int tid, int64 tick, int id, intptr_t data)
 {
 	if (login_fd > 0 && session[login_fd] != NULL)
 	{
@@ -4577,7 +4577,7 @@ int ping_login_server(int tid, unsigned int tick, int id, intptr_t data)
 //Invoked 15 seconds after mapif_disconnectplayer in case the map server doesn't
 //replies/disconnect the player we tried to kick. [Skotlex]
 //------------------------------------------------
-static int chardb_waiting_disconnect(int tid, unsigned int tick, int id, intptr_t data)
+static int chardb_waiting_disconnect(int tid, int64 tick, int id, intptr_t data)
 {
 	struct online_char_data* character;
 	if ((character = (struct online_char_data*)idb_get(online_char_db, id)) != NULL && character->waiting_disconnect == tid)
@@ -4601,7 +4601,7 @@ static int online_data_cleanup_sub(DBKey key, void *data, va_list ap)
 	return 0;
 }
 
-static int online_data_cleanup(int tid, unsigned int tick, int id, intptr_t data)
+static int online_data_cleanup(int tid, int64 tick, int id, intptr_t data)
 {
 	online_char_db->foreach(online_char_db, online_data_cleanup_sub);
 	return 0;
