+++ src/map/instance.c
@@ -340,7 +340,7 @@ void instance_destroy_freesvar(void *key, void *data, va_list args)
 /*--------------------------------------
  * Timer to destroy instance by process or idle
  *--------------------------------------*/
-int instance_destroy_timer(int tid, unsigned int tick, int id, intptr_t data)
+int instance_destroy_timer(int tid, int64 tick, int id, intptr_t data)
 {
 	instance_destroy(id);
 	return 0;
